+++ lib/l10n/index.dart
@@ -1,0 +1,1 @@
+export 'package:flutter_gen/gen_l10n/gallery_localizations.dart';