+++ lib/presentation/screens/fullscreen/video/video_fullscreen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_gen/gen_l10n/gallery_localizations.dart';
+import 'package:gallery/l10n/index.dart';
 import 'package:gallery_service/gallery_service.dart';
 import 'package:layout_service/layout_service.dart';
 import 'package:video_player/video_player.dart';