+++ services/gallery_service/lib/src/bloc/src/event.dart
@@ -8,6 +8,15 @@ abstract class GalleryEvent extends Equatable {
   List<Object?> get props => [];
 }
 
+class GetGalleryRequested extends GalleryEvent {
+  const GetGalleryRequested(this.path);
+
+  final String path;
+
+  @override
+  List<Object?> get props => [path];
+}
+
 class GalleryPushRequested extends GalleryEvent {
   const GalleryPushRequested(this.path);