+++ services/routers_service/lib/routers_service.dart
@@ -1,0 +1,4 @@
+library routers_service;
+
+export 'src/models/index.dart';
+export 'src/routers/index.dart';