+++ services/routers_service/lib/src/routers/index.dart
@@ -1,3 +1,2 @@
-export 'src/delegate.dart';
 export 'src/routes.dart';
 export 'src/state.dart';